Move cache into base class
